refactor(agents): extract AgentSaveDelegate from AgentEditorViewModel
Move the persistence mutations into AgentSaveDelegate: form validation +
create/update (save), plus duplicate, delete, and revertToVersion. The delegate
assembles the request body from the UI state (delegating pure transforms to
AgentEditorMappers) and emits the SaveSuccess/DuplicateSuccess/DeleteSuccess
events. The VM exposes the same public methods as thin delegations.

AgentEditorViewModel is now a thin orchestrator (1938 -> ~540 lines) wiring six
delegates over a shared AgentEditorStateHandle.

No behavior change.
